+++ decidim-admin/spec/shared/manage_impersonations_examples.rb
@@ -39,7 +39,7 @@ shared_examples "manage impersonations examples" do
       let(:document_number) { "123456789Y" }
 
       it "shows the errors in the form" do
-        expect(page).to have_selector("label", text: "Document number* is invalid")
+        expect(page).to have_selector("label", text: "Document number*\nis invalid")
       end
     end