+++ decidim-core/lib/decidim/form_builder.rb
@@ -229,6 +229,7 @@ module Decidim
       picker_options[:class] += " is-invalid-input" if error?(attribute)
 
       items = object.send(attribute).collect { |item| [item, yield(item)] }
+
       template = ""
       template += label(attribute, label_for(attribute) + required_for_attribute(attribute)) unless options[:label] == false
       template += @template.render("decidim/widgets/data_picker", picker_options: picker_options, prompt_params: prompt_params, items: items)