+++ decidim-proposals/app/cells/decidim/proposals/proposal_m_cell.rb
@@ -7,6 +7,7 @@ module Decidim
     # This cell renders a proposal with its M-size card.
     class ProposalMCell < Decidim::Proposals::ProposalCell
       include Cell::ViewModel::Partial
+      include Decidim::TooltipHelper
 
       def show
         render
@@ -20,11 +21,23 @@ module Decidim
         render
       end
 
+      def status
+        render
+      end
+
       def badge
         render
       end
 
       private
+
+      def current_user
+        context[:current_user]
+      end
+
+      def decidim
+        Decidim::Core::Engine.routes.url_helpers
+      end
 
       def resource_path
         resource_locator(model).path
@@ -32,6 +45,16 @@ module Decidim
 
       def body
         truncate(model.body, length: 100)
+      end
+
+      def badge_classes
+        classes = [badge_state_css]
+        classes += if options[:full_badge]
+                     ["label", "proposal-status"]
+                   else
+                     ["card__text--status"]
+                   end
+        classes.join(" ")
       end
     end
   end